COMPLETE

+++ 1feblab2021/details.php
@@ -1,13 +1,14 @@
 <?php
 include('config.php');
 ?>
-
 <?php
 $sql= "SELECT * FROM users";
 $sql1= "SELECT * FROM student_details";
 $result=mysqli_query($conn,$sql);
 $result1=mysqli_query($conn,$sql1);
-
+echo "<font size='6'>";
+echo nl2br("\nUsers Table Data");
+            echo "</font>";
 if($result->num_rows >0){?>
     <table border="1px">
         <thead>
@@ -24,7 +25,6 @@ if($result->num_rows >0){?>
         </thead>
         <tbody>
         <?php
-        // while($row=$result->fetch_array())
         while($row=$result->fetch_assoc() and $row1=$result1->fetch_assoc()){?>
             <tr>
                 <td><?php echo $row['username']?></td>
@@ -35,7 +35,7 @@ if($result->num_rows >0){?>
                 <td><?php echo $row1['year']?></td>
                 <td><a href="edit.php?id=<?php echo $row['id']?>">
                 <input type="button" value="Edit"></a></td>
-                <td><a href="delete.php?id=<?php echo $row['id']?>">
+                <td><a href="delete.php?id=<?php echo $row['id']?>" onclick="return confirm('Are you sure you want to delete?')">
                 <input type="button" value="Delete"></a></td>
             </tr>
             <?php } ?>
